+++ tests/testthat/test-to_dummy.R
@@ -1,0 +1,10 @@
+context("sjmisc, count_na")
+
+library(sjmisc)
+
+x <- to_dummy(data.frame(x = c("yes", "no", "yes", "maybe"), stringsAsFactors = F), suffix = "label")
+y <- to_dummy(data.frame(x = c("yes", "no", "yes", "maybe")), suffix = "label")
+
+test_that("to_dummy", {
+  expect_identical(x, y)
+})